Revert no-void rule change
We should work on making `parse` return one type consistently; this is
just a temporary change until we figure out if it should actually be
a promise.

The main strength of the floating-promises rule is to avoid errors
getting thrown off the stack. A good way to solve the problem is by
adding a basic `catch` clause to the function call, e.g.:

this.parse().catch(console.error)